Fixed a bug where SqlDataReader.GetSchemaTable() would throw an error with UDT queries.
